refactor: Apply code review feedback from previous PR
