+++ README.md
@@ -9,9 +9,14 @@
 * [텍스트와 이미지로 살펴보는 온라인 코드 리뷰 과정](https://github.com/next-step/nextstep-docs/tree/master/codereview)
 
 ## [ 자동차 경주 ]
-### 기능 구현 목록
+### 기능 구현 목록 - 자동차 경주
 - [x] 사용자에게 자동차 대수와 시도할 횟수에 대한 input 값을 입력받는다.
 - [x] 입력한 대수만큼 자동차를 생성한다.
 - [x] 0 ~ 9 사이의 random 값을 생성한다.
 - [x] 랜덤 숫자가 4 이상이면 한 칸 전진하며, 입력한 횟수만큼 반복한다.
 - [x] 자동차의 상태를 화면에 출력한다.
+
+### 기능 구현 목록 - 자동차 경주 (우승자)
+- [ ] 자동차의 수가 아닌 이름을 입력받게 하며, 자동차 이름은 5자를 초과할 수 없다.
+- [ ] 자동차의 상태를 이름과 함께 출력한다.
+- [ ] 경주가 종료되면 우승자를 출력한다.
