+++ src/main/java/study/racinggame/domain/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package study.step3.domain.strategy;
+package study.racinggame.domain.strategy;
 
 public interface MoveStrategy {
     boolean canMove();
