+++ src/main/java/study/racinggame/domain/strategy/RandomMoveStrategy.java
@@ -1,4 +1,4 @@
-package study.step3.domain.strategy;
+package study.racinggame.domain.strategy;
 
 import java.util.Random;
